Move Stripe billing and checkout session functionality to the actions folder
