+++ app/billing/page.jsx
@@ -6,8 +6,8 @@ import Header from "@/components/Header";
 import { useSession } from "next-auth/react";
 // import { useState } from "react";
 import { FaCreditCard } from "react-icons/fa";
-import { CreateBillingSession } from "@/app/api/stripe/billing-session/route";
-import { CancelSubscription } from "@/app/api/stripe/subscription-cancel/route";
+import { CreateBillingSession } from "@/actions/stripe/billing-session";
+import { CancelSubscription } from "@/actions/stripe/subscription-cancel";
 import { useRouter } from "next/navigation";
 import { useState, useMemo, useCallback, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
